pyproject.toml
[project]
name = "cagi-indel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]
